requirements.txt
boto3
pandas==2.2.2
requests==2.31.0
python-dotenv==1.0.1
tqdm==4.64.1